requirements.txt
docutils
#django-autoslug
dcg-autoslug

django-markuphelpers
django-webcal
django-dept-directory
